+++ tests/src/test/java/de/sovity/edc/e2e/DataSourceParameterizationTest.java
@@ -110,7 +110,6 @@ class DataSourceParameterizationTest {
     private final String destinationPath = "/destination/some/path/";
     private final String sourceUrl = "http://localhost:" + port + sourcePath;
     private final String destinationUrl = "http://localhost:" + port + destinationPath;
-    // TODO: remove the test backend dependency?
     private ClientAndServer mockServer;
 
     private static final AtomicInteger DATA_OFFER_INDEX = new AtomicInteger(0);
@@ -502,7 +501,7 @@ class DataSourceParameterizationTest {
         Map<String, String> dataSinkProperties = new HashMap<>();
         dataSinkProperties.put(EDC_NAMESPACE + "baseUrl", destinationUrl);
         dataSinkProperties.put(EDC_NAMESPACE + "method", HttpMethod.PUT);
-        dataSinkProperties.put(EDC_NAMESPACE + "type", "HttpData"); // TODO: http proxy
+        dataSinkProperties.put(EDC_NAMESPACE + "type", "HttpData");
         transferProcessProperties.put(rootKey + METHOD, testCase.method);
 
         if (testCase.body != null) {