+++ utils/json-and-jsonld-utils/src/main/java/de/sovity/edc/utils/jsonld/vocab/Prop.java
@@ -136,6 +136,16 @@ public class Prop {
         }
     }
 
+    @UtilityClass
+    public class SovityMessageExt {
+        public final String CTX = "https://semantic.sovity.io/message/generic/";
+        public final String REQUEST = CTX + "request";
+        public final String RESPONSE = CTX + "response";
+        public final String ERROR_MESSAGE = CTX + "errorMessage";
+        public final String HEADER = CTX + "header";
+        public final String BODY = CTX + "body";
+    }
+
     /**
      * FOAF Vocabulary
      */